requirements.txt
matplotlib==3.5.2
nibabel==4.0.2
numpy==1.23.3
opencv_python==4.6.0.66
pydicom==2.3.0
scikit_image==0.19.3
scipy==1.9.1
# torch==1.13.0
torchsummary==1.5.1
# torchvision==0.14.0
tqdm==4.64.1
